+++ AutoPkgr/LGDefaults.h
@@ -27,6 +27,17 @@ typedef NS_ENUM(NSInteger, LGApplicationDisplayStyle) {
     kLGDisplayStyleShowDock = 1 << 1,
 };
 
+typedef NS_ENUM(NSInteger, LGReportItems) {
+    kLGReportItemsNone = 0,
+    kLGReportItemsAll = 1 << 0,
+    kLGReportItemsNewDownloads = 2 << 0,
+    kLGReportItemsNewPackages = 2 << 1,
+    kLGReportItemsNewInstalls = 2 << 2,
+    kLGReportItemsMunkiImports = 2 << 3,
+    kLGReportItemsJSSImports = 2 << 4,
+};
+
+
 @interface LGDefaults : NSUserDefaults
 
 #pragma mark - Singletons
@@ -44,6 +55,11 @@ typedef NS_ENUM(NSInteger, LGApplicationDisplayStyle) {
  *  You can pass in either an NSDate object, or NSString and it will return a formatted date string
  */
 @property (copy, nonatomic) id LastAutoPkgRun;
+
+/**
+ *  Binary shifted values representing the keys to include in the emailed report
+ */
+@property (assign, nonatomic) LGReportItems reportedItemFlags;
 
 #pragma mark - BOOL
 @property (nonatomic) BOOL SMTPTLSEnabled;